Move skills

+++ script.sh
@@ -1,6 +1,8 @@
 for f in *; do
     if [[ -d $f ]]; then
-        pip install -r "$f/requirements.txt"
+        if [ -e "$f/requirements.txt" ]; then
+            pip install -r "$f/requirements.txt"
+        fi
         if [ "$f" != "mycroft-core" ]; then
             cp -r $f mycroft-core/mycroft/skills/$f
         fi
